+++ packages/core-ethereum/src/ethereum.mock.ts
@@ -1,10 +1,10 @@
 import { type ChainOptions } from './index.js'
 
 export const sampleChainOptions: ChainOptions = {
-  chainId: 1337,
-  environment: 'hardhat-localhost',
+  chainId: 31337,
+  environment: 'anvil-localhost',
   maxFeePerGas: '10 gwei',
   maxPriorityFeePerGas: '1 gwei',
-  network: 'hardhat',
+  network: 'anvil',
   provider: 'http://localhost:8545'
 }