+++ scripts/testnet.sh
@@ -174,7 +174,7 @@ fund_if_empty() {
 }
 
 # $1=vm name
-# Run a VM with a hardhat instance
+# Run a VM with an anvil instance
 start_chain_provider(){
   gcloud compute instances create-with-container $1-provider $GCLOUD_DEFAULTS \
       --create-disk name=$(disk_name $1),size=10GB,type=pd-standard,mode=rw \
@@ -193,19 +193,18 @@ add_keys() {
   fi
 }
 
-# $1 hardhat debug log file
-start_local_hardhat() {
+# $1 anvil debug log file
+start_local_anvil() {
   # Remove previous log file to make sure that the regex does not match
-  rm -f "${hardhat_rpc_log}"
+  log "remove ${1}"
+  rm -f "${1}"
 
-  log "Running hardhat local node"
-  HOPR_ENVIRONMENT_ID="hardhat-localhost" \
-    TS_NODE_PROJECT="$(yarn workspace @hoprnet/hopr-ethereum exec pwd)/tsconfig.hardhat.json" \
-    NODE_OPTIONS="--experimental-wasm-modules" \
-    yarn workspace @hoprnet/hopr-ethereum hardhat node \
-      --network hardhat \
-      --show-stack-traces > \
-      "$1" 2>&1 &
+  log "Running anvil local node at ${mydir}"
+  # make -C "${mydir}/.." run-anvil > \
+  #     "$1" 2>&1 &
+  # FIXME: stucking at here...
+  anvil > "$1" 2>&1 &
+  make -C "${mydir}/../packages/ethereum/contracts/" -j anvil-deploy-all 2>&1 &
 }
 
 # $1 prefix, e.g. "e2e-source"
@@ -219,27 +218,21 @@ fund_nodes() {
   local addr_arg=""
   [[ -n "${4:-}" ]] && addr_arg="--address ${4}"
 
-  HOPR_ENVIRONMENT_ID=hardhat-localhost \
-  TS_NODE_PROJECT="$(yarn workspace @hoprnet/hopr-ethereum exec pwd)/tsconfig.hardhat.json" \
-  NODE_OPTIONS="--experimental-wasm-modules" \
-    yarn workspace @hoprnet/hopr-ethereum hardhat faucet \
-      --identity-prefix "${node_prefix}" \
-      --identity-directory "${tmp}" \
-      --use-local-identities \
-      --network hardhat \
-      --password "${password}" \
-      ${addr_arg}
+  foundry-tool --environment-name anvil-localhost --environment-type development \
+    faucet --password "${password}" --use-local-identities \
+    --identity-prefix "${node_prefix}" --identity-directory "${tmp}" \
+    --private-key 0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80 \
+    --make-root "../packages/ethereum/contracts" \
+    ${addr_arg}
 }
 
 
 disable_network_registry() {
   log "Disabling register"
-  HOPR_ENVIRONMENT_ID=hardhat-localhost \
-  TS_NODE_PROJECT="$(yarn workspace @hoprnet/hopr-ethereum exec pwd)/tsconfig.hardhat.json" \
-  NODE_OPTIONS="--experimental-wasm-modules" \
-  yarn workspace @hoprnet/hopr-ethereum hardhat register \
-    --network hardhat \
-    --task disable
+  PRIVATE_KEY=0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80 \
+  make -C "../" disable-network-registry \
+  environment=anvil-localhost \
+  environment_type=development
 
   log "Register disabled"
 }