+++ test/integration-test.sh
@@ -100,22 +100,11 @@ redeem_tickets() {
   [[ ${redeemed} -gt 0 && ${redeemed} -gt ${last_redeemed} ]] || { msg "redeemed tickets count on node ${node_id} is ${redeemed}, previously ${last_redeemed}"; exit 1; }
 }
 
-# Performs a hardhat-specific EVM RPC call
-# to disable auto-mining at runtime
-disable_hardhat_auto_mining() {
-  log "Disabling hardhat automining"
-  env \
-    HOPR_ENVIRONMENT_ID=hardhat-localhost \
-    TS_NODE_PROJECT="./tsconfig.hardhat.json" \
-    yarn workspace @hoprnet/hopr-ethereum hardhat disable-automine \
-      --network hardhat
-}
-
 # $1 native addresses ("Ethereum addresses"), comma-separated list
 # $2 peerIds, comma-separated list
 register_nodes() {
   make -C "${mydir}/.. register-nodes" \
-    environment=hardhat-localhost \
+    environment=anvil-localhost environment_type=development \
     native_addresses="${1}" \
     peer_ids="${2}"
 }
@@ -123,20 +112,15 @@ register_nodes() {
 enable_network_registry() {
   log "Enabling register"
   env \
-    HOPR_ENVIRONMENT_ID=hardhat-localhost \
-    TS_NODE_PROJECT="./tsconfig.hardhat.json" \
-    yarn workspace @hoprnet/hopr-ethereum hardhat register \
-      --network hardhat \
-      --task enable
+    make -C "${mydir}/.. enable-network-registry" \
+    environment=anvil-localhost environment_type=development
 
   log "Register enabled"
 }
 
 log "Running full E2E test with ${api1}, ${api2}, ${api3}, ${api4}, ${api5}, ${api6}, ${api7}"
 
-# Setup is done, so disable hardhat's auto-mining to correctly mimic
 # real blockchain networks
-disable_hardhat_auto_mining
 
 validate_native_address "${api1}" "${api_token}" & jobs+=( "$!" )
 validate_native_address "${api2}" "${api_token}" & jobs+=( "$!" )
@@ -206,22 +190,13 @@ if ! [ -z $additional_nodes_addrs ] && ! [ -z $additional_nodes_peerids ]; then
 fi
 
 # Register nodes in the NR, emit "Registered" events
-HOPR_ENVIRONMENT_ID=hardhat-localhost \
-TS_NODE_PROJECT=${mydir}/../packages/ethereum/tsconfig.hardhat.json \
-yarn workspace @hoprnet/hopr-ethereum hardhat register \
-  --network hardhat \
-  --task add \
-  --native-addresses "${native_addrs_to_register}" \
-  --peer-ids "${native_peerids_to_register}"
+register_nodes "${native_addrs_to_register}" "${native_peerids_to_register}"
 log "Nodes added to register (Registered)"
 
 # Sync nodes in the NR, emit "EligibilityUpdated" events
-HOPR_ENVIRONMENT_ID=hardhat-localhost \
-TS_NODE_PROJECT=${mydir}/../packages/ethereum/tsconfig.hardhat.json \
-yarn workspace @hoprnet/hopr-ethereum hardhat register \
-  --network hardhat \
-  --task sync \
-  --peer-ids "${native_peerids_to_register}"
+make -C "${mydir}/.. sync-eligibility" \
+    environment=anvil-localhost environment_type=development \
+    peer_ids="${native_peerids_to_register}"
 log "Nodes added to register (EligibilityUpdated)"
 
 # running withdraw and checking it results at the end of this test run